[feature] (WIP) JDTEntryClass
This feature aim to use the oriented object paradigm in the projet.

The JDTEntryClass is just a simple object that manage a start_time,
end_time and a body. Some basic methods to access thoses are present and
one can call the get_md_entry() method to get the MarkDown result of the
entry.

Additionally, thoses JDTEntry objects are added to a ArrayList.

There is still things to manage, such as:
  * Modification of an entry
  * Sorting of the entries

JDTEntries should also probably be a Class to manage modification, sorting and
export, in whatever format needed — MarkDown, PDF, JSON, ...